Framework: Adding the possibility to lock the editor

+++ editor/components/block-list/block.js
@@ -3,7 +3,7 @@
  */
 import { connect } from 'react-redux';
 import classnames from 'classnames';
-import { get, partial, reduce, size } from 'lodash';
+import { get, partial, reduce, size, flow } from 'lodash';
 
 /**
  * WordPress dependencies
@@ -13,6 +13,7 @@ import { keycodes } from '@wordpress/utils';
 import { getBlockType, BlockEdit, getBlockDefaultClassname, createBlock, hasBlockSupport } from '@wordpress/blocks';
 import { withFilters } from '@wordpress/components';
 import { __, sprintf } from '@wordpress/i18n';
+import { withContext } from '@wordpress/components';
 
 /**
  * Internal dependencies
@@ -294,7 +295,7 @@ class BlockListBlock extends Component {
 			case ENTER:
 				// Insert default block after current block if enter and event
 				// not already handled by descendant.
-				if ( target === this.node ) {
+				if ( target === this.node && ! this.props.isLocked ) {
 					event.preventDefault();
 
 					this.props.onInsertBlocks( [
@@ -316,12 +317,14 @@ class BlockListBlock extends Component {
 			case DELETE:
 				// Remove block on backspace.
 				if ( target === this.node ) {
+					const { uid, onRemove, previousBlock, onFocus, isLocked } = this.props;
 					event.preventDefault();
-					const { uid, onRemove, previousBlock, onFocus } = this.props;
-					onRemove( uid );
+					if ( ! isLocked ) {
+						onRemove( uid );
 
-					if ( previousBlock ) {
-						onFocus( previousBlock.uid, { offset: -1 } );
+						if ( previousBlock ) {
+							onFocus( previousBlock.uid, { offset: -1 } );
+						}
 					}
 				}
 				break;
@@ -338,7 +341,7 @@ class BlockListBlock extends Component {
 	}
 
 	render() {
-		const { block, order, mode, showContextualToolbar } = this.props;
+		const { block, order, mode, showContextualToolbar, isLocked } = this.props;
 		const { name: blockName, isValid } = block;
 		const blockType = getBlockType( blockName );
 		// translators: %s: Type of block (i.e. Text, Image etc)
@@ -408,10 +411,10 @@ class BlockListBlock extends Component {
 								focus={ focus }
 								attributes={ block.attributes }
 								setAttributes={ this.setAttributes }
-								insertBlocksAfter={ this.insertBlocksAfter }
-								onReplace={ onReplace }
+								insertBlocksAfter={ isLocked ? undefined : this.insertBlocksAfter }
+								onReplace={ isLocked ? undefined : onReplace }
 								setFocus={ partial( onFocus, block.uid ) }
-								mergeBlocks={ this.mergeBlocks }
+								mergeBlocks={ isLocked ? undefined : this.mergeBlocks }
 								className={ className }
 								id={ block.uid }
 							/>
@@ -514,7 +517,14 @@ const mapDispatchToProps = ( dispatch, ownProps ) => ( {
 	},
 } );
 
-export default compose(
+export default flow(
 	withFilters( 'Editor.BlockItem' ),
-	connect( mapStateToProps, mapDispatchToProps )
+	connect( mapStateToProps, mapDispatchToProps ),
+	withContext( 'editor' )( ( settings ) => {
+		const { templateLock } = settings;
+
+		return {
+			isLocked: true === templateLock,
+		};
+	} ),
 )( BlockListBlock );
